// Teryt/TerytFrontend/src/app/shared/responsiveness/responsives.enum.ts
export enum OrientationType {
  Unknown = 'Unknown',
  Portrait = 'Portrait',
  Landscape = 'Landscape'
}

export enum DeviceType {
  Unknown = 'Unknown',
  Tablet = 'Tablet',
  Handset = 'Handset',
  Web = 'Web'
}

export enum ScreenSizeType {
  Unknown = 'Unknown',
  XSmall = 'XSmall',
  Small = 'Small',
  Medium = 'Medium',
  Large = 'Large',
  XLarge = 'XLarge'
}

export enum BreakpointType {
  Unknown = 'Unknown',
  HandsetPortrait = 'HandsetPortrait',
  TabletPortrait = 'TabletPortrait',
  TabletLandscape = 'TabletLandscare',
  WebPortrait = 'WebPortrait',
  WebLandscape = 'WebLandscape'
}